pyproject.toml
[project]
name = "finance-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "watchdog>=6.0.0",
]
